+++ src/Controller/TourController.php
@@ -24,9 +24,12 @@ class TourController extends AbstractController
     /**
      * @Route("/", name="tour_index")
      */
-    public function index(TourRepository $repo)
+    public function index()
     {
-        $tours = $repo->findAll();
+        $permissions = $this->getUser()->getPermissions();
+        foreach($permissions as $p){
+            $tours[] = $p->getTour();
+        }
         return $this->render('tour/index.html.twig', [
             'tours' => $tours,
         ]);
@@ -40,9 +43,15 @@ class TourController extends AbstractController
      */
     public function show(Tour $tour): Response
     {
-        return $this->render('tour/show.html.twig', [
-            'tour' => $tour
-        ]);
+        $permissions = $tour->getPermissions();
+        foreach ($permissions as $p){
+            if($p->getUser() === $this->getUser()){
+                return $this->render('tour/show.html.twig', [
+                    'tour' => $tour
+                ]);
+            }
+        }
+        return $this->redirectToRoute("home_index");
     }
 
     /**
@@ -83,28 +92,34 @@ class TourController extends AbstractController
      */
     public function edit(Request $request, Tour $tour): Response
     {
-        $form = $this->createForm(TourType::class, $tour);
-        $form->handleRequest($request);
+        $permissions = $tour->getPermissions();
+        foreach ($permissions as $p) {
+            if ($p->getUser() === $this->getUser()) {
+                $form = $this->createForm(TourType::class, $tour);
+                $form->handleRequest($request);
 
-        if($form->isSubmitted()){
-            if($form->isValid()){
-                $this->getDoctrine()->getManager()->flush();
+                if ($form->isSubmitted()) {
+                    if ($form->isValid()) {
+                        $this->getDoctrine()->getManager()->flush();
 
-                //TODO ajouter un message de validation d'edition
-                return $this->redirectToRoute('tour_show', [
-                    'id' => $tour->getId()
+                        //TODO ajouter un message de validation d'edition
+                        return $this->redirectToRoute('tour_show', [
+                            'id' => $tour->getId()
+                        ]);
+                    }
+                    //TODO ajouter un message d'erreur de formulaire non valide
+                }
+                return $this->render('tour/edit.html.twig', [
+                    'tour' => $tour,
+                    'form' => $form->createView(),
                 ]);
             }
-            //TODO ajouter un message d'erreur de formulaire non valide
         }
-        return $this->render('tour/edit.html.twig', [
-            'tour' => $tour,
-            'form' => $form->createView(),
-        ]);
+        return $this->redirectToRoute("home_index");
     }
 
     /**
-     * @param EntityManagerInterface $em
+     *
      * @param Tour $tour
      * @return Response
      *
@@ -112,11 +127,16 @@ class TourController extends AbstractController
      */
     public function delete(Tour $tour)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($tour);
-        $entityManager->flush();
-
-        //TODO ajouter un message de confirmation de suppression
-        return $this->redirectToRoute('tour_index');
+        $permissions = $tour->getPermissions();
+        foreach ($permissions as $p) {
+            if ($p->getUser() === $this->getUser()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($tour);
+                $entityManager->flush();
+                //TODO ajouter un message de confirmation de suppression
+                return $this->redirectToRoute('tour_index');
+            }
+        }
+        return $this->redirectToRoute("home_index");
     }
 }
